Login added which first uses RSA encryption to check username against password.
Then server sends a nonce thats encrypted using RSA.

User then loads certificate and also signs the nonce to authenticate them-selves and sends to the server.

Server verifies certificate and grabs user public key from it and uses it to verify the signature and then verifies if the nonce is valid.

Server generates session key and encrypts it using the users public key.

Then using the session encrypts user credentials and sends them over.

User decrypt this with the session key and is now logged in.